Add REPL and other changes
Moved definitions out of __init__.py since __init__.py is not loaded
when __main__.py is run. Definition of run() is now in run.py. Added a
basic repl() function for a REPL. Add __main__.py to run the REPL or a
script. Finally, moved __version__ into _version.py for the same reason
as moving run().